+++ src/components/spartan/SAccordion/SAccordion.vue
@@ -10,30 +10,39 @@ const props = defineProps<{
 
 <template>
     <Transition appear :name="`${vertical ? 'vertical' : 'horizontal'}-accordion`">
-        <div v-show="open" :class="twMerge('overflow-hidden', vertical ? 'h-full' : 'w-full', props.class)">
+        <div
+            v-show="open"
+            :class="
+                twMerge(vertical ? 'max-h-screen overflow-y-hidden' : 'max-w-screen overflow-x-hidden', props.class)
+            "
+        >
             <slot />
         </div>
     </Transition>
 </template>
 
 <style scoped>
+.max-w-screen {
+    max-width: 100vw;
+}
+
 .horizontal-accordion-enter-active,
 .horizontal-accordion-leave-active {
-    transition: width 0.4s cubic-bezier(0.87, 0, 0.13, 1);
+    transition: max-width 0.4s cubic-bezier(0.87, 0, 0.13, 1);
 }
 
 .horizontal-accordion-enter-from,
 .horizontal-accordion-leave-to {
-    width: 0;
+    max-width: 0;
 }
 
 .vertical-accordion-enter-active,
 .vertical-accordion-leave-active {
-    transition: height 0.4s cubic-bezier(0.87, 0, 0.13, 1);
+    transition: max-height 0.4s cubic-bezier(0.87, 0, 0.13, 1);
 }
 
 .vertical-accordion-enter-from,
 .vertical-accordion-leave-to {
-    height: 0;
+    max-height: 0;
 }
 </style>