+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 2.0.0-beta.32 - 2023-11-07
 ### Added
 - STable context and auto generation modes
 - `STableHead`, `STableBody`, `STableRow`, `STableCell`,  and `STableHeadCell` components implementation.